% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbox.R
\name{gbox}
\alias{gbox}
\title{Create a quick box plot in ggplot.}
\usage{
gbox(x, y = NULL, horizontal = FALSE, xlab = "x", ylab = "y", title = "title")
}
\arguments{
\item{x}{A numeric vector for which box plot needs to be plotted. If a second argument is provided, this should be the categorical variable.}

\item{y}{A categorical vector if side-by-side box plots are needed (Optional).}

\item{horizontal}{Logical indicating whether the box plot(s) should be horizontal (TRUE) or vertical (FALSE).}

\item{xlab}{This is the label for the x-axis.}

\item{ylab}{This is the label for the y-axis.}

\item{title}{This is the title for the histogram.}
}
\value{
This function returns a ggplot box plot object.
}
\description{
This function creates a ggplot box plot of a numeric vector. If a numeric
and a categorical variable are given, it creates side-by-side box plots.
The x-axis will be labeled "x", the y-axis labeled "y", and the title
labeled "title" by default.
}
\examples{
## Create a box plot(s) of a single (or multiple) variable(s)
x <- rnorm(100)
gbox(x)
gbox(x, sample(c("A", "B"), 100, replace = TRUE))

}
